refactor(clawdstrike): extract merge from policy.rs
Pure relocation of all extends/merge logic into a sibling module:
- impl GuardConfigs { merge_with } (the deep guard-config merge)
- merge_prompt_injection_config, merge_jailbreak_config (presence-aware helpers)
- impl Policy { merge } (the 2nd Policy impl block; legal cross-module split)
- merge_custom_guards

merge.rs imports merge_verification_settings via the mod.rs pub(crate) re-export.
Test-only ForbiddenPathConfig/SecretLeakConfig imports moved into the inline
test module; the four config types only the merge code used dropped from mod.rs.
No logic edits.
